added factory test
