+++ server/.dockerignore
@@ -1,2 +1,3 @@
 npm-debug.log
 .dockerignore
+coverage
